Moved some simple and useful metrics from maf/mafContrib to maf/metrics
